+++ app/celery/LLM/tasks.py
@@ -1,6 +1,6 @@
 # app/tasks.py
 
-from .celery_app import celery_app
+from ..celery_app import celery_app
 
 @celery_app.task
 def example_task(x, y):
