+++ app/services/web_crawling/community_website/dcinside.py
@@ -7,7 +7,7 @@ from concurrent.futures import ThreadPoolExecutor, as_completed
 
 from constants import DEFAULT_GPT_ANSWER, SITE_DCINSIDE
 from db.mongo_controller import MongoController
-from services.web_crwaling.community_website.community_website import AbstractCommunityWebsite
+from services.web_crawling.community_website.community_website import AbstractCommunityWebsite
 from utils import FTPClient
 from config import Config
