+++ app/services/web_crawling/community_website/instiz.py
@@ -2,7 +2,7 @@ import re
 from bs4 import BeautifulSoup
 import requests
 from db.mongo_controller import MongoController
-from services.web_crwaling.community_website.community_website import AbstractCommunityWebsite
+from services.web_crawling.community_website.community_website import AbstractCommunityWebsite
 from datetime import datetime
 from utils import FTPClient
 import logging
