+++ app/services/web_crawling/index.py
@@ -7,12 +7,12 @@ from fastapi.middleware.cors import CORSMiddleware
 from pymongo import settings
 
 from db.mongo_controller import MongoController
-from services.web_crwaling.community_website.instiz import Instiz
-from services.web_crwaling.community_website.ppomppu import Ppomppu
-from services.web_crwaling.community_website.ruliweb import Ruliweb
-from services.web_crwaling.community_website.theqoo import Theqoo
-from services.web_crwaling.community_website.ygosu import Ygosu
-from services.web_crwaling.community_website.dcinside import Dcinside
+from services.web_crawling.community_website.instiz import Instiz
+from services.web_crawling.community_website.ppomppu import Ppomppu
+from services.web_crawling.community_website.ruliweb import Ruliweb
+from services.web_crawling.community_website.theqoo import Theqoo
+from services.web_crawling.community_website.ygosu import Ygosu
+from services.web_crawling.community_website.dcinside import Dcinside
 
 from utils.llm import LLM
 from utils.tag_split import Tagsplit
